refactored to remove incentives API call Now derived from prisoner search

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/digitalcanteenapi/client/btPinPhoneClient/dto/BtPinPhoneClientDto.kt
@@ -3,6 +3,6 @@ package uk.gov.justice.digital.hmpps.digitalcanteenapi.client.btPinPhoneClient.d
 data class BtPinPhoneClientDto(
   val reference: String,
   val prisonerId: String,
-  val balancePence: Double,
-  val creditLimitPounds: Double,
+  val balancePence: Int,
+  val creditLimitPounds: Int,
 )
